rcfpch is updated to read all data from the connection. Compilation warnings are fixed
git-svn-id: https://svn.oktetlabs.ru/svnroot/prj/oktetlabs/test-environment/src/trunk@15403 9d14d920-75d5-0310-8bd7-a52572981aee

+++ lib/rcfpch/rcf_pch_conf.c
@@ -574,7 +574,8 @@ process_wildcard(struct rcf_comm_connection *conn, char *cbuf,
         SEND_ANSWER("%d", TE_RC(TE_RCF_PCH, rc));
         
     if ((size_t)snprintf(cbuf + answer_plen, buflen - answer_plen,
-                         "0 attach %u", strlen(tmp) + 1) >=
+                         "0 attach %u", 
+                         (unsigned int)(strlen(tmp) + 1)) >=
             (buflen - answer_plen))
     {
         free(tmp);
